refactor(ui): rehome foundation utilities
