+++ npm/ui/core/vite.config.ts
@@ -157,8 +157,8 @@ export default defineConfig({
       checkbox: "src/families/selection/checkbox/checkbox.ts",
       collection: "src/collection.ts",
       "composite-navigation": "src/composite-navigation.ts",
-      context: "src/context.ts",
-      "controllable-state": "src/controllable-state.ts",
+      context: "src/families/foundations/context/context.ts",
+      "controllable-state": "src/families/foundations/controllable-state/controllable-state.ts",
       dialog: "src/families/overlays/dialog/dialog.ts",
       "alert-dialog": "src/families/overlays/alert-dialog/alert-dialog.ts",
       "dismissable-layer": "src/dismissable-layer.ts",
